add delete button and create job button

+++ src/main/java/de/tum/cit/aet/job/service/JobService.java
@@ -91,6 +91,9 @@ public class JobService {
      * @param jobId the ID of the job to delete
      */
     public void deleteJob(UUID jobId) {
+        if (!jobRepository.existsById(jobId)) {
+            throw new EntityNotFoundException("Job id " + jobId + " does not exist");
+        }
         jobRepository.deleteById(jobId);
     }
